Spun off SurveyPolygon.ts
