+++ src/web/command_control/client/components/SurveyLines.ts
